Feat#52 - 엔티티 PK int -> Long으로 수정

+++ neulbomgil-backend/src/main/java/com/gooroomees/neulbomgil_backend/domain/auth/entity/RefreshToken.java
@@ -14,12 +14,12 @@ public class RefreshToken {
     private Long id;
 
     @Column(nullable = false, unique = true)
-    private int userId;
+    private Long userId;
 
     @Column(nullable = false)
     private String refreshToken;
 
-    public RefreshToken(int userId, String refreshToken) {
+    public RefreshToken(Long userId, String refreshToken) {
         this.userId = userId;
         this.refreshToken = refreshToken;
     }
